Change string-merging to use a concurrent hashmap
This change was an experiment with the goal of improving performance, however
it actually made performance significantly worse. Increasing the time required
to link a debug build of clang from about 9 seconds to about 17. While there is
some room for improvement, the main problem seems to be when deduplicating the
strings by accessing the concurrent hashmap.